overhauled the julia simulation code
All the particle data is now contained inside the `Colloid` struct
and the polygon classes are removed.

There is now a `Simulation` struct that acts as the main container
for the simulation data and the `Colloid` is now used
for handling particle data only.

Each Monte Carlo step now updates every particle in the simulation
(either by a translation or a rotation).

Validating the moves is now optimized with a cell list to avoid doing
extra overlap checks.

The new code is written with future GPU functionallity in mind